refactor common fixtures for better reuse.  added fixture to test EMPRO staff users.
